create class Main.java. refactor class User.java  - added constructor, @Setter, @Getter, @ToString,@EqualsAndHashCode  - rename adress to address. delete UserRegister.java.
