--- src/main/java/faang/school/godbless/UserRegister.java
@@ -1,2 +1,0 @@
-package faang.school.godbless;public class UserRegister {
-}
