+++ src/main/java/faang/school/godbless/groupbyage/User.java
@@ -1,20 +1,28 @@
-package faang.school.godbless;
+package faang.school.godbless.groupbyage;
 
 import lombok.*;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
 
-@AllArgsConstructor
-@Data
+@Setter
+@Getter
+@ToString
+@EqualsAndHashCode
 public class User {
     private String name;
     private int age;
     private String job;
-    private String adress;
+    private String address;
+
+    public User(String name, int age, String job, String address) {
+        this.name = name;
+        this.age = age;
+        this.job = job;
+        this.address = address;
+    }
 
     public static Map<Integer, List<User>> groupUsers(List<User> users) {
         Map<Integer, List<User>> groupUsers = new TreeMap<>();
@@ -26,4 +34,5 @@ public class User {
         }
         return groupUsers;
     }
+
 }
